requirements.txt
jiwer
pymupdf
pdfminer.six
python-docx
pytesseract
Pillow
tika
pandas
pytest
black

nltk
spacy

requests
python-dotenv